ID.ENH: add first version of EPU50 window
