+++ pyqt-apps/siriushla/si_id_control/__init__.py
@@ -1,2 +1,5 @@
-from .apu_control import APUControlWindow
+"""ID controls."""
+
+from .apu import APUControlWindow
+from .epu import EPUControlWindow
 from .id_control import IDControl
